+++ goalc/compiler/StaticObject.cpp
@@ -16,17 +16,11 @@ uint32_t push_data_to_byte_vector(T data, std::vector<uint8_t>& v) {
 ////////////////
 // StaticString
 ////////////////
-StaticString::StaticString(std::string data, int _seg) : text(std::move(data)), seg(_seg) {}
+StaticString::StaticString(std::string text_data, int _seg)
+    : StaticBasic(_seg, "string"), text(std::move(text_data)) {}
 
 std::string StaticString::print() const {
   return fmt::format("static-string \"{}\"", text);
-}
-
-StaticObject::LoadInfo StaticString::get_load_info() const {
-  LoadInfo info;
-  info.requires_load = false;
-  info.prefer_xmm = false;
-  return info;
 }
 
 void StaticString::generate(emitter::ObjectGenerator* gen) {
@@ -47,10 +41,6 @@ void StaticString::generate(emitter::ObjectGenerator* gen) {
     d.push_back(c);
   }
   d.push_back(0);
-}
-
-int StaticString::get_addr_offset() const {
-  return BASIC_OFFSET;
 }
 
 ////////////////
@@ -110,6 +100,14 @@ void StaticStructure::generate_structure(emitter::ObjectGenerator* gen) {
   for (auto& sym : symbols) {
     gen->link_static_symbol_ptr(rec, sym.offset, sym.name);
   }
+
+  for (auto& ptr : pointers) {
+    gen->link_static_pointer(rec, ptr.offset_in_this, ptr.dest->rec, ptr.offset_in_dest);
+  }
+
+  for (auto& type : types) {
+    gen->link_static_type_ptr(rec, type.offset, type.name);
+  }
 }
 
 void StaticStructure::generate(emitter::ObjectGenerator* gen) {
@@ -123,18 +121,94 @@ void StaticStructure::add_symbol_record(std::string name, int offset) {
   symbols.push_back(srec);
 }
 
+void StaticStructure::add_pointer_record(int offset_in_this,
+                                         StaticStructure* dest,
+                                         int offset_in_dest) {
+  PointerRecord prec;
+  prec.offset_in_this = offset_in_this;
+  prec.dest = dest;
+  prec.offset_in_dest = offset_in_dest;
+  pointers.push_back(prec);
+}
+
+void StaticStructure::add_type_record(std::string name, int offset) {
+  SymbolRecord srec;
+  srec.name = std::move(name);
+  srec.offset = offset;
+  types.push_back(srec);
+}
+
 ///////////////////
 // StaticBasic
 ///////////////////
 
 StaticBasic::StaticBasic(int _seg, std::string _type_name)
-    : StaticStructure(_seg), type_name(std::move(_type_name)) {}
+    : StaticStructure(_seg), type_name(std::move(_type_name)) {
+  add_type_record(type_name, 0);
+}
 
 int StaticBasic::get_addr_offset() const {
   return BASIC_OFFSET;
 }
 
-void StaticBasic::generate(emitter::ObjectGenerator* gen) {
+///////////////////
+// StaticPair
+///////////////////
+
+StaticPair::StaticPair(StaticResult car, StaticResult cdr, int _seg)
+    : StaticStructure(_seg), m_car(std::move(car)), m_cdr(std::move(cdr)) {}
+
+int StaticPair::get_addr_offset() const {
+  return PAIR_OFFSET;
+}
+
+void StaticPair::generate(emitter::ObjectGenerator* gen) {
+  data.resize(2 * POINTER_SIZE);  // size of pair
+  generate_item(m_car, 0);
+  generate_item(m_cdr, 4);
   generate_structure(gen);
-  gen->link_static_type_ptr(rec, 0, type_name);
+}
+
+void StaticPair::generate_item(const StaticResult& item, int offset) {
+  if (item.is_reference()) {
+    add_pointer_record(offset, item.reference(), item.reference()->get_addr_offset());
+  } else if (item.is_symbol()) {
+    add_symbol_record(item.symbol_name(), offset);
+    u32 symbol_placeholder = 0xffffffff;
+    memcpy(data.data() + offset, &symbol_placeholder, POINTER_SIZE);
+  } else if (item.is_constant_data()) {
+    // if it's a constant data, it should always be a boxed integer for a pair.
+    // or I guess you could put a normal integer too. Either way, we assume signed here,
+    // though we may need to allow overflow so you can store either signed/unsigned things in pairs
+    s32 value = item.get_as_s32();
+    memcpy(data.data() + offset, &value, POINTER_SIZE);
+  }
+}
+
+///////////////////
+// StaticResult
+///////////////////
+
+StaticResult StaticResult::make_structure_reference(StaticStructure* structure, TypeSpec ts) {
+  StaticResult result;
+  result.m_kind = Kind::STRUCTURE_REFERENCE;
+  result.m_struct = structure;
+  result.m_ts = std::move(ts);
+  return result;
+}
+
+StaticResult StaticResult::make_constant_data(u64 value, TypeSpec ts) {
+  StaticResult result;
+  result.m_kind = Kind::CONSTANT_DATA;
+  result.m_constant_data = value;
+  result.m_ts = std::move(ts);
+  return result;
+}
+
+StaticResult StaticResult::make_symbol(const std::string& name) {
+  StaticResult result;
+  result.m_kind = Kind::SYMBOL;
+  result.m_symbol = name;
+  result.m_ts = TypeSpec("symbol");
+  return result;
 }